feat: redo permissioning for security & efficiency

+++ apps/api/app/user/user_service.ts
@@ -1,9 +1,8 @@
-import assert from 'node:assert/strict';
 import type { HttpContext } from '@adonisjs/core/http';
-import { TRPCError } from '@trpc/server';
 import { eq } from 'drizzle-orm';
 import * as Schema from '#database/schema';
 import type { AppBouncer } from '#middleware/initialize_bouncer_middleware';
+import { AuthorizationService } from '../authorization/authorization_service.js';
 import { db } from '../db/db_service.js';
 import type { UserSchema } from './schemas/user_schema.js';
 import { UserTimezoneSchema } from './schemas/user_timezone_schema.js';
@@ -13,7 +12,7 @@ export class UserService {
 		bouncer: AppBouncer,
 		user: Pick<UserSchema, 'id'>,
 	): Promise<Pick<UserSchema, 'displayName'> | undefined> {
-		assert(await bouncer.with('UserPolicy').allows('read', user), new TRPCError({ code: 'FORBIDDEN' }));
+		await AuthorizationService.assertPermission(bouncer.with('UserPolicy').allows('read', user));
 
 		const dbUser = await db.query.users.findFirst({
 			where: eq(Schema.users.id, user.id),
@@ -26,7 +25,7 @@ export class UserService {
 	}
 
 	async deleteUser(bouncer: AppBouncer, user: Pick<UserSchema, 'id'>): Promise<void> {
-		assert(await bouncer.with('UserPolicy').allows('delete', user), new TRPCError({ code: 'FORBIDDEN' }));
+		await AuthorizationService.assertPermission(bouncer.with('UserPolicy').allows('delete', user));
 
 		await db.delete(Schema.users).where(eq(Schema.users.id, user.id));
 	}
@@ -36,7 +35,7 @@ export class UserService {
 		user: Pick<UserSchema, 'id'>,
 		updated: Pick<UserSchema, 'displayName'>,
 	): Promise<void> {
-		assert(await bouncer.with('UserPolicy').allows('update', user), new TRPCError({ code: 'FORBIDDEN' }));
+		await AuthorizationService.assertPermission(bouncer.with('UserPolicy').allows('update', user));
 
 		await db.update(Schema.users).set(updated).where(eq(Schema.users.id, user.id));
 	}
